Move k8s client config into internal package
